refactor : 패키지 묶음 계층형 구조로 변경
